terraform/cloud-functions/datastore-refresher/requirements.txt
google-cloud-discoveryengine==0.13.5
google-api-python-client==2.156.0
google-cloud-aiplatform==1.106.0
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.31.1
python-dotenv==1.1.1